+++ Security/src/main/java/de/muenchen/security/configurator/JDBCAuthenticationConfigurator.java
@@ -1,4 +1,4 @@
-package de.muenchen.security;
+package de.muenchen.security.configurator;
 
 import org.kohsuke.MetaInfServices;
 import org.springframework.beans.factory.annotation.Autowired;
