+++ Security/src/main/java/de/muenchen/security/entities/AuthPermId.java
@@ -1,5 +1,8 @@
 
-package de.muenchen.security;
+package de.muenchen.security.entities;
+
+import de.muenchen.security.entities.Authority;
+import de.muenchen.security.entities.Permission;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
