+++ Security/src/main/java/de/muenchen/security/repositories/AuthorityPermissionRepository.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package de.muenchen.security;
+package de.muenchen.security.repositories;
 
+import de.muenchen.security.entities.AuthPermId;
+import de.muenchen.security.entities.AuthorityPermission;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
