+++ Security/src/main/java/de/muenchen/security/repositories/AuthorityRepository.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package de.muenchen.security;
+package de.muenchen.security.repositories;
 
+import de.muenchen.security.entities.Authority;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
@@ -18,21 +19,23 @@ import java.util.List;
  * @author praktikant.tmar
  */
 @RepositoryRestResource(exported = false)
-public interface PermissionRepository  extends CrudRepository<Permission, Long>  {
-    
-    public final static String Permission_CACHE = "PERMISSION_CACHE";
+public interface AuthorityRepository  extends CrudRepository<Authority, Long> {
 
-    @Cacheable(value = Permission_CACHE, key = "#p0")
-    public Permission findFirstByOid(String oid);
+    public final static String Authority_CACHE = "AUTHORITYPERMISSION_CACHE";
 
-    @Override
-    @CachePut(value = Permission_CACHE, key = "#p0.oid")
-    public Permission save(Permission entity);
+    @Cacheable(value = Authority_CACHE, key = "#p0")
+    public Authority findFirstByOid(String oid);
 
     @Override
-    @CacheEvict(value = Permission_CACHE, key = "#p0.oid")
-    public void delete(Permission entity);
+    @CachePut(value = Authority_CACHE, key = "#p0.oid")
+    public Authority save(Authority entity);
 
-    public List<Permission> findByOid(String oid);
+
+    @Override
+    @CacheEvict(value = Authority_CACHE, key = "#p0.oid")
+    public void delete(Authority entity);
+
+    public List<Authority> findByOid(String oid);
+
 }
