+++ vaadin-demo-micro-service/src/main/java/de/muenchen/demo/service/domain/CompanyBaseInfo.java
@@ -5,8 +5,6 @@
  */
 package de.muenchen.demo.service.domain;
 
-import java.util.HashSet;
-import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -14,7 +12,7 @@ import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
-import de.muenchen.security.SecurityEntity;
+import de.muenchen.security.entities.SecurityEntity;
 import org.hibernate.search.annotations.Indexed;
 
 /**
