+++ vaadin-demo-micro-service/src/main/java/de/muenchen/demo/service/domain/Sachbearbeiter.java
@@ -1,7 +1,7 @@
 package de.muenchen.demo.service.domain;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import de.muenchen.security.User;
+import de.muenchen.security.entities.User;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
